Refactor Kippy API into modular package
